Finished lab and technical documentation
I unfortunately used a VM that uses an older version of Git Bash. This meant that instead of pushing to main, I ended up pushing to a new branch titled 'master'. I'm adding a note to the 'no' file to basically say ignore it. This was added using touch so that I could make a commit containing something and push it when troubleshooting.